% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classdefs.R
\name{Spc.AsPolylines}
\alias{Spc.AsPolylines}
\title{Outputs all object coordinates as a list of 3 column matrices corresponding
to (x,y,z) coordinates of relevant points. Treats spheres and planes as special cases,
since these obviously can't be accurately reproduced in wire-frame form}
\usage{
Spc.AsPolylines(obj)
}
\arguments{
\item{object}{The object (elementary or compound) to be plotted}
}
\value{
List of matrices
}
\description{
Outputs all object coordinates as a list of 3 column matrices corresponding
to (x,y,z) coordinates of relevant points. Treats spheres and planes as special cases,
since these obviously can't be accurately reproduced in wire-frame form
}
\examples{
  Spc.AsPolylines (my_world)
}
\seealso{
Other outputs: \code{\link{Spc.Plot}}
}
