% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compound.R
\name{Spc.Combine}
\alias{Spc.Combine}
\title{Return an object from a list of other objects}
\usage{
Spc.Combine(objects, properties = NA, bound = TRUE)
}
\arguments{
\item{objects}{A list of sub-objects}

\item{properties}{Package-independent object defining additional properties.
Default NA}

\item{bound}{Should a bounding sphere be added to the new object? Default = TRUE}
}
\value{
Returns a compound object consisting of the objects passed to it.
  The function constructs an object and assigns it the necessary properties
  for the package to recognise it as such.
}
\description{
Return an object from a list of other objects
}
\examples{
  world <- Spc.Combine(list(my_table, my_chairs), "dining area", TRUE)
}
\seealso{
Other constructors: \code{\link{Spc.MakeCuboid}},
  \code{\link{Spc.MakePlane}},
  \code{\link{Spc.MakeSphere}},
  \code{\link{Spc.MakeTriangle}}
}
