% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classdefs.R
\name{Spc.Intersect}
\alias{Spc.Intersect}
\title{Find the distance at which a ray first intersects with an object, plus any relevant object
properties at the point of intersection}
\usage{
Spc.Intersect(ray.origin, ray.direction, object)
}
\arguments{
\item{ray.origin}{A positional vector of where the ray originates}

\item{ray.direction}{A directional vector for the ray's direction}

\item{object}{The object (elementary or compound) we're testing for intersection}
}
\value{
A list of intersect properties, consisting of:
   distance = intersect distance (as proportin of the length of ray.direction);
   properties = the *properties* attribute from the intersected object (or NA if none)
}
\description{
Find the distance at which a ray first intersects with an object, plus any relevant object
properties at the point of intersection
}
\examples{
  Spc.Intersect (c(0,0,0), c(0,1,0),  my_world)
}
