% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuboid.R
\name{Spc.MakeCuboid}
\alias{Spc.MakeCuboid}
\title{Return a cuboid object, constructed from elementary spacial objects}
\usage{
Spc.MakeCuboid(centre, dimensions, properties = NA, bound = TRUE)
}
\arguments{
\item{centre}{Centre of cuboid}

\item{dimensions}{Basic dimensions of the cuboid, as a vector of length 3
in (x,y,z) dimensions.}

\item{properties}{Package-independent object defining additional cuboid properties}
}
\value{
A compound cuboid object, comprising of elementary triangle objects
  in a grouped hierarchical object, with a surrounding bounding sphere
  used for spacial indexing.
}
\description{
Return a cuboid object, constructed from elementary spacial objects
}
\examples{
  w <- Spc.MakeCuboid(c(0,0,0), c(2,2,2), surface_props)
}
\seealso{
Other constructors: \code{\link{Spc.CombineObjects}},
  \code{\link{Spc.MakePlane}},
  \code{\link{Spc.MakeSphere}},
  \code{\link{Spc.MakeTriangle}}
}
