% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plane.R
\name{Spc.MakePlane}
\alias{Spc.MakePlane}
\title{Return an elementary plane object}
\usage{
Spc.MakePlane(point, normal, properties = NA, direction.north = NA,
  direction.east = NA)
}
\arguments{
\item{point}{A point on the plane}

\item{normal}{A vector which is normal to the direction of the plane}

\item{properties}{Package-independent object defining additional plane properties.
Default NA}
}
\value{
Elementary plane object.
}
\description{
Return an elementary plane object
}
\examples{
  pl <- Spc.MakePlane(c(0,0,0), c(0,1,0), "Dave")
}
\seealso{
Other constructors: \code{\link{Spc.Combine}},
  \code{\link{Spc.MakeCuboid}},
  \code{\link{Spc.MakePolygon}},
  \code{\link{Spc.MakeSphere}},
  \code{\link{Spc.MakeTriangle}}
}
