% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{Spc.MakeSphere}
\alias{Spc.MakeSphere}
\title{Return an elementary sphere object}
\usage{
Spc.MakeSphere(centre, radius, properties = NA)
}
\arguments{
\item{centre}{Centre of sphere}

\item{radius}{of sphere}

\item{properties}{Package-independent object defining additional sphere properties
Default NA}
}
\value{
Elementary sphere object
}
\description{
Return an elementary sphere object
}
\examples{
  s <- Spc.MakeSphere(c(0,0,0), 2, surface_props)
}
\seealso{
Other constructors: \code{\link{Spc.Combine}},
  \code{\link{Spc.MakeCuboid}},
  \code{\link{Spc.MakePlane}},
  \code{\link{Spc.MakeTriangle}}
}
