% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle.R
\name{Spc.MakeTriangle}
\alias{Spc.MakeTriangle}
\title{Return an elementary triangle object}
\usage{
Spc.MakeTriangle(A, B, C, properties = NA)
}
\arguments{
\item{A, B, C}{A, B, and C define the corners of the triangle in space}

\item{properties}{Package-independent object defining additional sphere properties
Default NA}
}
\value{
Elementary triangle object. The function does not check that the 
  3 input points are not colinear. If this is the case, the result won't be much
  of a triangle, and will simply define a zero area line.
}
\description{
Return an elementary triangle object
}
\examples{
  t <- Spc.MakeTriangle(c(0,0,0), c(1,1,1), c(2,0,10), surface_props)
}
\seealso{
Other constructors: \code{\link{Spc.MakeCuboid}},
  \code{\link{Spc.MakeSphere}}
}
