% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Spc.Plot}
\alias{Spc.Plot}
\title{Plot an orthographic plan view along x, y and z axes}
\usage{
Spc.Plot(object)
}
\arguments{
\item{object}{The object (elementary or compound) to be plotted}
}
\description{
Plot an orthographic plan view along x, y and z axes
}
\examples{
  Spc.Plot (my_world)
}
\seealso{
Other outputs: \code{\link{Spc.AsPolylines}}
}
