% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classdefs.R
\name{Spc.Rotate}
\alias{Spc.Rotate}
\title{Rotate an object by c(ax,ay,az) in the x y and z planes, about the point c(x,y,z) 
Uses the right-hand coordinate system (this is the only functin in this package which is handed in this way).
Rotates about first the x axis, then y, then z. This operation is not commutable (i.e. if you rotate by the various
angles about first z, then x then y, you'd get a different result).}
\usage{
Spc.Rotate(pivot.point = c(0, 0, 0), pivot.angle, object)
}
\arguments{
\item{pivot.point}{Vector of the point in space to rotate about. Defaults to c(0,0,0)}

\item{pivot.angle}{Vector of angles to rotate about (x,y,z) - in degrees}

\item{object}{The object (elementary or compound) to be rotated}
}
\value{
The same object, but with all appropriate spacial points updated accordingly
}
\description{
Rotate an object by c(ax,ay,az) in the x y and z planes, about the point c(x,y,z) 
Uses the right-hand coordinate system (this is the only functin in this package which is handed in this way).
Rotates about first the x axis, then y, then z. This operation is not commutable (i.e. if you rotate by the various
angles about first z, then x then y, you'd get a different result).
}
\examples{
  my_object <- Spc.Rotate(c(0,0,0),c(45,0,0),my_object)
}
\seealso{
Other transforms: \code{\link{Spc.Translate}}
}
