% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classdefs.R
\name{Spc.Translate}
\alias{Spc.Translate}
\title{Move the object hierarchy to a different point in space, without performing rotation or any other trsnformation}
\usage{
Spc.Translate(object, move.direction)
}
\arguments{
\item{object}{The object (elementary or compound) to be moved}

\item{move.direction}{A directional vector containing the distance (x,y,z) by which the object should be moved}
}
\value{
The same object, but with all appropriate spacial points updated accordingly
}
\description{
Move the object hierarchy to a different point in space, without performing rotation or any other trsnformation
}
\examples{
  my_object <- Spc.Translate (my_object, c(0,20,0))
}
\seealso{
Other transforms: \code{\link{Spc.Rotate}}
}
