% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Utils.CrossProduct}
\alias{Utils.CrossProduct}
\title{Cross Product of two direction vectors in 3D space}
\usage{
Utils.CrossProduct(a, b)
}
\arguments{
\item{a}{First direction vector}

\item{b}{Second direction vector}
}
\value{
Three element vector which is perpentical to the plane containing
  the two input vectors. Returns c(0,0,0) is the two inputs are parallel
  or if either are zero length.

 Note: Utils.CrossProduct(a, b) == - Utils.CrossProduct(b, a)

 This function is specifically coded for 3D space, i.e. totally expects
 both a and b to be 3 element vectors
}
\description{
Cross Product of two direction vectors in 3D space
}
\examples{
  Utils.CrossProduct(c(3,2,1),c(1,2,3))  # [1]  4 -8  4
  Utils.CrossProduct(c(1,1,1),c(2,2,2))  # [1] 0 0 0

}
\references{
\url{https://en.wikipedia.org/wiki/Cross_product}
}
\seealso{
Other utils: \code{\link{Utils.DotProduct}},
  \code{\link{Utils.UnitVector}},
  \code{\link{Utils.VectorLength}}
}
