% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Utils.DotProduct}
\alias{Utils.DotProduct}
\title{Dot (Scalar) Product of two direction vectors in 3D space}
\usage{
Utils.DotProduct(a, b)
}
\arguments{
\item{a}{First direction vector}

\item{b}{Second direction vector}
}
\value{
Scalar value for dot rpdocut.
  Returns zero if the two inputs are perpendicular
  or if either are of zero length.
}
\description{
Dot (Scalar) Product of two direction vectors in 3D space
}
\examples{
  Utils.DotProduct(c(3,2,1),c(1,2,3))  # [1] 10
  Utils.DotProduct(c(1,1,0),c(1,-1,0)) # [1] 0

}
\references{
\url{https://en.wikipedia.org/wiki/Dot_product}
}
\seealso{
Other utils: \code{\link{Utils.CrossProduct}},
  \code{\link{Utils.UnitVector}},
  \code{\link{Utils.VectorLength}}
}
