% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Utils.VectorLength}
\alias{Utils.VectorLength}
\title{The scalar length of a vector}
\usage{
Utils.VectorLength(a)
}
\arguments{
\item{a}{The vector}
}
\value{
Scalar length - think Pythagorus in 3D.
  Will "work" on vectors of any number of dimensions, although it should
  be noted it's designed for 3D space use.
}
\description{
The scalar length of a vector
}
\examples{
  Utils.VectorLength(c(1,2,3))  # [1] 3.741657
  Utils.VectorLength(c(3,4,0))  # [1] 5
}
\seealso{
Other utils: \code{\link{Utils.CrossProduct}},
  \code{\link{Utils.DotProduct}},
  \code{\link{Utils.UnitVector}}
}
