% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plane.R
\name{Spc.MakePlane}
\alias{Spc.MakePlane}
\title{Return an elementary plane object}
\usage{
Spc.MakePlane(point, normal, properties = NA)
}
\arguments{
\item{point}{A point on the plane}

\item{normal}{A vector which is normal to the direction of the plane}

\item{properties}{Package-independent object defining additional plane properties.
Default NA}
}
\value{
Elementary plane object. The function neither checks that the 
  3 input points are not colinear, nor that any of the 3 points are equal.
  Expect odd-ish results in either of those cases, since the resulting
  object will we either a line or a point with zero area.
}
\description{
Return an elementary plane object
}
\examples{
  pl <- Spc.MakePlane(c(0,0,0), c(0,1,0), "Dave")
}
\seealso{
Other constructors: \code{\link{Spc.Combine}},
  \code{\link{Spc.MakeCuboid}},
  \code{\link{Spc.MakePolygon}},
  \code{\link{Spc.MakeSphere}},
  \code{\link{Spc.MakeTriangle}}
}
