% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{Spc.MakeSphere}
\alias{Spc.MakeSphere}
\title{Return an elementary sphere object}
\usage{
Spc.MakeSphere(centre, radius, properties = NA, direction.pole = NA,
  direction.meridian = NA)
}
\arguments{
\item{centre}{Centre of sphere}

\item{radius}{of sphere}

\item{properties}{Package-independent object defining additional sphere properties
Default NA}

\item{direction.pole, direction.meridian}{Nominal north pole direction vector, and a direction vector 
to a point on the prime meridian.
If supplied, an intersect will also calculate and return components $north and $east, as the
polar coordinates north (-90 to 90) and east (0 - 360) of the intersection point on the sphere.
Alse, where supplied, these must not be parallel.}
}
\value{
Elementary sphere object
}
\description{
Return an elementary sphere object
}
\examples{
  s <- Spc.MakeSphere(c(0,0,0), 2, surface_props)
}
\seealso{
Other constructors: \code{\link{Spc.Combine}},
  \code{\link{Spc.MakeCuboid}},
  \code{\link{Spc.MakePlane}},
  \code{\link{Spc.MakePolygon}},
  \code{\link{Spc.MakeTriangle}}
}
