% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Spc.Plot}
\alias{Spc.Plot}
\title{Plot an orthographic plan view along x, y and z axes}
\usage{
Spc.Plot(object, view.axis = NA)
}
\arguments{
\item{object}{The object (elementary or compound) to be plotted}

\item{view.axis}{Set to "X","Y" or "Z" to only plot the view along that axis.
Otherwise will plot all 3 views together.}
}
\description{
Plot an orthographic plan view along x, y and z axes
}
\examples{
  Spc.Plot (my_world)
}
\seealso{
Other outputs: \code{\link{Spc.AsPolylines}}
}
