% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classdefs.R
\name{Spc.Translate}
\alias{Spc.Translate}
\title{Move the object hierarchy to a different point in space, without performing rotation or any other trsnformation}
\usage{
Spc.Translate(move.direction, object)
}
\arguments{
\item{move.direction}{A directional vector containing the distance (x,y,z) by which the object should be moved}

\item{object}{The object (elementary or compound) to be moved}
}
\value{
The same object, but with all appropriate spacial points updated accordingly
}
\description{
Move the object hierarchy to a different point in space, without performing rotation or any other trsnformation
}
\examples{
  my_object <- Spc.Translate (c(0,20,0), my_object)
}
